Invert the responsibility of writing data in WriteListener
It used to be that the `WriteController` would simply notify the `Http2StreamState`
that it could write and the stream would be in charge of putting data into the controller.
That relationship is now inverted in that the `WriteController` simply asks for data and
adds it to the output buffer itself.